requirements.txt
flask==2.3.3
flask-cors==4.0.0
langchain==0.1.4
langchain-groq==0.0.1
python-dotenv==1.0.0
gTTS==2.5.1
SpeechRecognition==3.10.1
pyttsx3==2.90
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-JWT-Extended==4.6.0
email-validator==2.1.0.post1
SQLAlchemy==2.0.25
